refactor: combine ndk service repository with crash processor
