+++ embrace-android-features/src/test/java/io/embrace/android/embracesdk/internal/ndk/NativeCrashProcessorImplTest.kt
@@ -4,7 +4,6 @@ import io.embrace.android.embracesdk.ResourceReader
 import io.embrace.android.embracesdk.concurrency.BlockableExecutorService
 import io.embrace.android.embracesdk.fakes.FakeConfigService
 import io.embrace.android.embracesdk.fakes.FakeJniDelegate
-import io.embrace.android.embracesdk.fakes.FakeNdkServiceRepository
 import io.embrace.android.embracesdk.fakes.FakeSharedObjectLoader
 import io.embrace.android.embracesdk.fakes.FakeStorageService
 import io.embrace.android.embracesdk.fakes.FakeSymbolService
@@ -24,46 +23,49 @@ import java.io.File
 internal class NativeCrashProcessorImplTest {
 
     private lateinit var service: NativeCrashProcessor
-    private lateinit var storageManager: FakeStorageService
     private lateinit var configService: FakeConfigService
     private lateinit var sharedObjectLoader: FakeSharedObjectLoader
     private lateinit var logger: EmbLogger
     private lateinit var delegate: FakeJniDelegate
-    private lateinit var repository: FakeNdkServiceRepository
     private lateinit var blockableExecutorService: BlockableExecutorService
+    private lateinit var storageService: FakeStorageService
+    private lateinit var ndkDir: File
     private val serializer = EmbraceSerializer()
 
     @Before
     fun setup() {
-        storageManager = FakeStorageService()
+        storageService = FakeStorageService()
         configService = FakeConfigService(autoDataCaptureBehavior = FakeAutoDataCaptureBehavior(ndkEnabled = true))
         sharedObjectLoader = FakeSharedObjectLoader().apply { loadEmbraceNative() }
         logger = EmbLoggerImpl()
         delegate = FakeJniDelegate()
-        repository = FakeNdkServiceRepository()
         blockableExecutorService = BlockableExecutorService()
     }
 
     @After
     fun after() {
-        val cacheDir = File("${storageManager.cacheDirectory}/ndk")
+        val cacheDir = File("${storageService.cacheDirectory}/ndk")
         if (cacheDir.exists()) {
             cacheDir.delete()
         }
-        val filesDir = File("${storageManager.filesDirectory}/ndk")
+        val filesDir = File("${storageService.filesDirectory}/ndk")
         if (filesDir.exists()) {
             filesDir.delete()
         }
     }
 
     private fun initializeService() {
+        ndkDir = File("${storageService.filesDirectory}/ndk").apply {
+            mkdirs()
+        }
+
         service = NativeCrashProcessorImpl(
             sharedObjectLoader,
             logger,
-            repository,
             delegate,
             serializer,
-            FakeSymbolService(mapOf("symbol1" to "test"))
+            FakeSymbolService(mapOf("symbol1" to "test")),
+            storageService
         )
     }
 
@@ -76,16 +78,14 @@ internal class NativeCrashProcessorImplTest {
 
     @Test
     fun `test getLatestNativeCrash catches an exception if _getCrashReport returns an empty string`() {
-        repository.addCrashFiles(File.createTempFile("test", "test"))
         initializeService()
+        addCrashFiles("test")
         val crashData = service.getLatestNativeCrash()
         assertNull(crashData)
     }
 
     @Test
     fun `test getLatestNativeCrash catches an exception if _getCrashReport returns invalid json syntax`() {
-        repository.addCrashFiles(File.createTempFile("test", "test"))
-
         val json = "{\n" +
             "  \"sid\": [\n" +
             "    {\n" +
@@ -95,20 +95,18 @@ internal class NativeCrashProcessorImplTest {
         delegate.crashRaw = json
 
         initializeService()
+        addCrashFiles("test")
         val crashData = service.getLatestNativeCrash()
         assertNull(crashData)
     }
 
     @Test
     fun `test getLatestNativeCrash when a native crash was captured`() {
-        repository.addCrashFiles(
-            nativeCrashFile = File.createTempFile("test", "crash-test")
-        )
         delegate.crashRaw = getNativeCrashRaw()
-
         configService.appFramework = AppFramework.UNITY
 
         initializeService()
+        addCrashFiles("test")
 
         val result = checkNotNull(service.getLatestNativeCrash())
         with(result) {
@@ -129,8 +127,8 @@ internal class NativeCrashProcessorImplTest {
     fun `getNativeCrashes returns all the crashes in the repository and doesn't invoke delete`() {
         delegate.crashRaw = getNativeCrashRaw()
         initializeService()
-        repository.addCrashFiles(File.createTempFile("file1", "tmp"))
-        repository.addCrashFiles(File.createTempFile("file2", "temp"))
+        addCrashFiles("file1")
+        addCrashFiles("file2")
         assertEquals(2, service.getNativeCrashes().size)
         assertEquals(2, service.getNativeCrashes().size)
     }
@@ -139,10 +137,14 @@ internal class NativeCrashProcessorImplTest {
     fun `getLatestNativeCrash returns only one crash even if there are many and deletes them all`() {
         delegate.crashRaw = getNativeCrashRaw()
         initializeService()
-        repository.addCrashFiles(File.createTempFile("file1", "tmp"))
-        repository.addCrashFiles(File.createTempFile("file2", "temp"))
+        addCrashFiles("file1")
+        addCrashFiles("file2")
         assertNotNull(service.getLatestNativeCrash())
         assertEquals(0, service.getNativeCrashes().size)
+    }
+
+    private fun addCrashFiles(name: String) {
+        File(ndkDir, "emb_ndk_$name.crash").createNewFile()
     }
 
     private fun getNativeCrashRaw() = ResourceReader.readResourceAsText("native_crash_raw.txt")
